share some common code for compile-time miri instances
